gestion_com-main/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
Django==5.1.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
Pillow==11.0.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.13.1
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
whitenoise==6.11.0
